dialects: (gate) move things to same file
