+++ inconspiquous/dialects/gate.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import ClassVar, Literal
+from typing import ClassVar, Literal, NamedTuple
 
 from xdsl.dialects.builtin import (
     AnyFloatConstr,
@@ -41,20 +41,23 @@ from xdsl.pattern_rewriter import RewritePattern
 from xdsl.printer import Printer
 from xdsl.traits import ConstantLike, Pure
 
-from inconspiquous.constraints import SizedAttributeConstraint
+from inconspiquous.constraints import SizedAttribute, SizedAttributeConstraint
 from inconspiquous.dialects.angle import AngleAttr, AngleType
-from inconspiquous.gates import (
-    GateAttr,
-    SingleQubitCliffordGate,
-    SingleQubitGate,
-    TwoQubitCliffordGate,
-)
-from inconspiquous.gates.core import (
-    CliffordGateAttr,
-    PauliGate,
-    PauliProp,
-    TwoQubitGate,
-)
+
+
+class GateAttr(ParametrizedAttribute, SizedAttribute, ABC):
+    """
+    In general most gate operations are not operationally different, so differentiating between them
+    may actually be better done via an attribute that can be attached to a gate operation.
+    """
+
+    @property
+    @abstractmethod
+    def num_qubits(self) -> int: ...
+
+    @property
+    def size(self) -> int:
+        return self.num_qubits
 
 
 @irdl_attr_definition
@@ -122,6 +125,94 @@ class ConstantGateOp(IRDLOperation, HasFolderInterface):
 
     def fold(self) -> tuple[GateAttr]:
         return (self.gate,)
+
+
+# Helper classes
+class SingleQubitGate(GateAttr):
+    @property
+    def num_qubits(self) -> int:
+        return 1
+
+
+class TwoQubitGate(GateAttr):
+    @property
+    def num_qubits(self) -> int:
+        return 2
+
+
+class PauliProp(NamedTuple):
+    """
+    Describes the combination of x and z pauli gates.
+    """
+
+    x: bool
+    z: bool
+
+    @staticmethod
+    def none() -> PauliProp:
+        return PauliProp(False, False)
+
+    @staticmethod
+    def X() -> PauliProp:
+        return PauliProp(True, False)
+
+    @staticmethod
+    def Y() -> PauliProp:
+        return PauliProp(True, True)
+
+    @staticmethod
+    def Z() -> PauliProp:
+        return PauliProp(False, True)
+
+    @staticmethod
+    def from_lit(literal: Literal["X", "Z"]) -> PauliProp:
+        if literal == "X":
+            return PauliProp.X()
+        return PauliProp.Z()
+
+
+class CliffordGateAttr(GateAttr, ABC):
+    """
+    Base class for Clifford gates that support Pauli propagation.
+    """
+
+    @abstractmethod
+    def pauli_prop(
+        self, input_idx: int, pauli_type: Literal["X", "Z"]
+    ) -> tuple[PauliProp, ...]:
+        """
+        Compute Pauli propagation through this gate.
+
+        Args:
+            input_idx: The index of the input qubit where the Pauli gate is applied
+            pauli_type: Either "X" or "Z" indicating the type of Pauli gate
+
+        Returns:
+            A PauliProp object where the `x` and `z` component determine whether
+            the corresponding Pauli component should be applied to that output.
+
+        For example, for Hadamard gate:
+            - X propagates to Z: pauli_prop(0, "X") returns ((x: False, z: True),)
+            - Z propagates to X: pauli_prop(0, "Z") returns ((x: True, z: False),)
+        """
+
+
+class SingleQubitCliffordGate(CliffordGateAttr, SingleQubitGate):
+    """Base class for single-qubit Clifford gates."""
+
+
+class PauliGate(SingleQubitCliffordGate):
+    """Base class for pauli gates"""
+
+    def pauli_prop(
+        self, input_idx: int, pauli_type: Literal["X", "Z"]
+    ) -> tuple[PauliProp, ...]:
+        assert input_idx == 0
+        return (PauliProp.from_lit(pauli_type),)
+
+
+class TwoQubitCliffordGate(CliffordGateAttr, TwoQubitGate):
+    """Base class for two-qubit Clifford gates."""
 
 
 @irdl_attr_definition
